feat : Article 조회수 업데이트, 수정 구현

+++ src/com/ssafy/ltw/domain/article/model/dao/ArticleDao.java
@@ -10,9 +10,9 @@ public interface ArticleDao {
     void writeArticle(ArticleDto articleDto) throws SQLException;
     List<ArticleDto> listArticle() throws SQLException;
     int getTotalArticleCount(Map<String, Object> param) throws SQLException;
-    ArticleDto getArticle(int id) throws SQLException;
-    void updateHit(int id) throws SQLException;
+    ArticleDto getArticle(Long id) throws SQLException;
+    void updateHit(Long id) throws SQLException;
 
     void modifyArticle(ArticleDto articleDto) throws SQLException;
-    void deleteArticle(int id) throws SQLException;
+    void deleteArticle(Long id) throws SQLException;
 }
